chamando calcula service para valores manuais

+++ AcompanhaAi/app/src/main/java/com/ai/acompanha/acompanhaai/service/ICalculaService.java
@@ -2,7 +2,7 @@ package com.ai.acompanha.acompanhaai.service;
 
 public interface ICalculaService  {
 
-    double calcularValor(int anterior, int atual);
+    float calcularValor(int anterior, int atual);
 
     int calculaConsumo(int anterior, int atual);
